+++ start_streams.bat
@@ -3,10 +3,10 @@ setlocal
 cd /d %~dp0
 set PYTHONIOENCODING=utf-8
 
-echo Iniciando streaming_unified.py (reinicio automático en caso de salida)...
+echo Iniciando streaming_combined.py (reinicio automático en caso de salida)...
 :loop
-python -u streaming_unified.py
+python -u streaming_combined.py
 set EXITCODE=%ERRORLEVEL%
-echo [%date% %time%] streaming_unified.py finalizó con código %EXITCODE%. Reinicio en 5 segundos...
+echo [%date% %time%] streaming_combined.py finalizó con código %EXITCODE%. Reinicio en 5 segundos...
 timeout /t 5 /nobreak >nul
 goto loop
